first steps towards a new full post view.
some weird bits in how we have to mangle primary and layout to get towards the view we want.
there's no support for suppressing the sidebar outside of a section configuration, and we can't make a section that's isolated to full post paths. The paths are only prefixes, not matches, so we can't say /read/feeds/:feed/posts, only read/feeds/, which isn't enough.

hacked in a thing to toggle a class on layout when the full post mounts and unmounts.